clean ibc dependencies
