+++ apps/src/lib/node/ledger/shell/finalize_block.rs
@@ -178,9 +178,7 @@ where
                         self.write_log.commit_tx();
                         tx_result["code"] = ErrorCodes::Ok.into();
                         if let Some(ibc_event) = &result.ibc_event {
-                            tx_result
-                                .merge_ibc_event(ibc_event)
-                                .map_err(Error::IbcEvent)?;
+                            tx_result.merge_ibc_event(ibc_event);
                         }
                         match serde_json::to_string(
                             &result.initialized_accounts,
