+++ apps/src/lib/node/ledger/shell/mod.rs
@@ -80,8 +80,6 @@ pub enum Error {
     Tendermint(tendermint_node::Error),
     #[error("Server error: {0}")]
     TowerServer(String),
-    #[error("Error converting an IBC event")]
-    IbcEvent(crate::node::ledger::events::Error),
 }
 
 /// The different error codes that the ledger may
