+++ tests/Cargo.toml
@@ -16,6 +16,7 @@ ABCI = [
   "ibc-proto-abci",
   "tendermint-stable",
   "anoma/ABCI",
+  "anoma/ibc-vp-abci",
   "anoma_vm_env/ABCI",
 ]
 
@@ -24,6 +25,7 @@ ABCI-plus-plus = [
   "ibc-proto",
   "tendermint",
   "anoma/ABCI-plus-plus",
+  "anoma/ibc-vp",
   "anoma_vm_env/ABCI-plus-plus",
 ]
