+++ vm_env/src/imports.rs
@@ -56,7 +56,7 @@ pub mod tx {
     use anoma::types::address;
     use anoma::types::address::Address;
     use anoma::types::chain::CHAIN_ID_LENGTH;
-    use anoma::types::ibc::IbcEvent;
+    use anoma::types::ibc::wrapper::IbcEvent;
     use anoma::types::internal::HostEnvResult;
     use anoma::types::storage::{
         BlockHash, BlockHeight, Epoch, BLOCK_HASH_LENGTH,
